pir support uniform_random
